+++ src/EulerHS/Core/Runtime.hs
@@ -3,14 +3,13 @@ module EulerHS.Core.Runtime
     -- * Core Runtime
     CoreRuntime(..)
   , LoggerRuntime(..)
-  , TransientLoggerContext
-  , setTransientContext
   , shouldLogRawSql
-  , clearTransientContext
   , incLogCounter
   , createCoreRuntime
   , createVoidLoggerRuntime
+  , createMemoryLoggerRuntime
   , createLoggerRuntime
+  , createLoggerRuntime'
   , clearCoreRuntime
   , clearLoggerRuntime
   , getLogMaskingConfig
@@ -18,42 +17,74 @@ module EulerHS.Core.Runtime
   ) where
 
 import           EulerHS.Prelude
-
-import qualified EulerHS.Core.Logger.ImplMimicPSBad.TinyLogger as Impl
 import           EulerHS.Core.Types
   ( LogCounter
   , LogLevel(..)
   , LoggerConfig(..)
-  , TransientLoggerContext
   , LogMaskingConfig(..)
   , ShouldLogSQL(SafelyOmitSqlLogs, UnsafeLogSQL_DO_NOT_USE_IN_PRODUCTION)
   )
+-- Currently, TinyLogger is highly coupled with the Runtime.
+-- Fix it if an interchangable implementations are needed.
+import qualified EulerHS.Core.Logger.Impl.TinyLogger as Impl
+import qualified EulerHS.Core.Types as T
 import           EulerHS.Core.Types.DB as X (withTransaction)
+import qualified System.Logger as Log
+
 
 -- TODO: add StaticLoggerRuntimeContext if we'll need more than a single Bool
 data LoggerRuntime
-  = LoggerRuntime !LogLevel !ShouldLogSQL !LogCounter TransientLoggerContext TransientLoggerContext (Maybe LogMaskingConfig) Impl.LoggerHandle
-  | MemoryLoggerRuntime !LogLevel !(MVar [Text])
+  = LoggerRuntime
+    { _flowFormatter    :: T.FlowFormatter
+    , _logLevel         :: T.LogLevel
+    , _logRawSql        :: ShouldLogSQL
+    , _logCounter       :: !T.LogCounter
+    , _logMaskingConfig :: Maybe T.LogMaskingConfig
+    , _logLoggerHandle  :: Impl.LoggerHandle
+    }
+  | MemoryLoggerRuntime !T.FlowFormatter !T.LogLevel !(MVar [Text]) !T.LogCounter
 
 data CoreRuntime = CoreRuntime
     { _loggerRuntime :: LoggerRuntime
     }
 
-createLoggerRuntime :: LoggerConfig -> IO LoggerRuntime
-createLoggerRuntime (MemoryLoggerConfig cfgLogLevel) =
-  MemoryLoggerRuntime cfgLogLevel <$> newMVar []
-createLoggerRuntime cfg = do
+-- createLoggerRuntime :: LoggerConfig -> IO LoggerRuntime
+-- createLoggerRuntime (MemoryLoggerConfig cfgLogLevel) =
+--   MemoryLoggerRuntime cfgLogLevel <$> newMVar []
+-- createLoggerRuntime cfg = do
+--   counter <- initLogCounter
+--   LoggerRuntime (_level cfg) (_logRawSql cfg) counter Nothing Nothing (_logMaskingConfig cfg)<$> Impl.createLogger cfg
+
+createMemoryLoggerRuntime :: T.FlowFormatter -> T.LogLevel -> IO LoggerRuntime
+createMemoryLoggerRuntime flowFormatter logLevel =
+  MemoryLoggerRuntime flowFormatter logLevel <$> newMVar [] <*> initLogCounter
+
+createLoggerRuntime :: T.FlowFormatter -> T.LoggerConfig -> IO LoggerRuntime
+createLoggerRuntime flowFormatter cfg = do
   counter <- initLogCounter
-  LoggerRuntime (_level cfg) (_logRawSql cfg) counter Nothing Nothing (_logMaskingConfig cfg)<$> Impl.createLogger cfg
+  LoggerRuntime flowFormatter (T._logLevel cfg) (T._logRawSql cfg) counter Nothing
+    <$> Impl.createLogger flowFormatter cfg
+
+createLoggerRuntime'
+  :: Maybe Log.DateFormat
+  -> Maybe Log.Renderer
+  -> T.BufferSize
+  -> T.FlowFormatter
+  -> T.LoggerConfig
+  -> IO LoggerRuntime
+createLoggerRuntime' mbDateFormat mbRenderer bufferSize flowFormatter cfg = do
+  counter <- initLogCounter
+  loggerHandle <- Impl.createLogger' mbDateFormat mbRenderer bufferSize flowFormatter cfg
+  pure $ LoggerRuntime flowFormatter (T._logLevel cfg) (T._logRawSql cfg) counter Nothing loggerHandle
 
 createVoidLoggerRuntime :: IO LoggerRuntime
 createVoidLoggerRuntime = do
   counter <- initLogCounter
-  LoggerRuntime Debug SafelyOmitSqlLogs counter Nothing Nothing Nothing <$> Impl.createVoidLogger
+  LoggerRuntime (const $ pure T.showingMessageFormatter) T.Debug SafelyOmitSqlLogs counter Nothing <$> Impl.createVoidLogger
 
 clearLoggerRuntime :: LoggerRuntime -> IO ()
-clearLoggerRuntime (LoggerRuntime _ _ _ _ _ _ handle) = Impl.disposeLogger handle
-clearLoggerRuntime _                              = pure ()
+clearLoggerRuntime (LoggerRuntime flowFormatter _ _ _ _ handle) = Impl.disposeLogger flowFormatter handle
+clearLoggerRuntime (MemoryLoggerRuntime _ _ msgsVar _) = void $ swapMVar msgsVar []
 
 createCoreRuntime :: LoggerRuntime -> IO CoreRuntime
 createCoreRuntime = pure . CoreRuntime
@@ -61,29 +92,18 @@ createCoreRuntime = pure . CoreRuntime
 clearCoreRuntime :: CoreRuntime -> IO ()
 clearCoreRuntime _ = pure ()
 
--- TODO: change Maybe Text to some type
-setTransientContext :: TransientLoggerContext -> TransientLoggerContext -> LoggerRuntime -> LoggerRuntime
-setTransientContext ctx1 ctx2 = \case
-  LoggerRuntime logLvl logRawSql counter _ _ maskConfig handle -> LoggerRuntime logLvl logRawSql counter ctx1 ctx2 maskConfig handle
-  rt -> rt
-
-clearTransientContext :: LoggerRuntime -> LoggerRuntime
-clearTransientContext (LoggerRuntime logLvl logRawSql counter _ _ _ handle) =
-  LoggerRuntime logLvl logRawSql counter Nothing Nothing Nothing handle
-clearTransientContext rt = rt
-
 shouldLogRawSql :: LoggerRuntime -> Bool
 shouldLogRawSql = \case
-  (LoggerRuntime _ UnsafeLogSQL_DO_NOT_USE_IN_PRODUCTION _ _ _ _ _) -> True
+  (LoggerRuntime _ _ UnsafeLogSQL_DO_NOT_USE_IN_PRODUCTION _ _ _) -> True
   _ -> False
 
-getLogMaskingConfig :: LoggerRuntime -> Maybe LogMaskingConfig
+getLogMaskingConfig :: LoggerRuntime -> Maybe T.LogMaskingConfig
 getLogMaskingConfig = \case
-  (LoggerRuntime _ _ _ _ _ mbMaskConfig _) -> mbMaskConfig
+  (LoggerRuntime _ _ _ _ mbMaskConfig _) -> mbMaskConfig
   _ -> Nothing
 
-initLogCounter :: IO LogCounter
+initLogCounter :: IO T.LogCounter
 initLogCounter = newIORef 0
 
-incLogCounter :: LogCounter -> IO Int
+incLogCounter :: T.LogCounter -> IO Int
 incLogCounter = flip atomicModifyIORef' (\cnt -> (cnt + 1, cnt))